chore: folders moved to backend
