+++ Backend/types/trade-data.ts
